// src/types/File/Pg/PgList.ts
import ItemList from '@/types/File/ItemList';

interface PgList {
    localTypeList: ItemList[];
    qualiteIsolationList: ItemList[];
    statutMenageTypeList: ItemList[];
    batimentNatureList: ItemList[];
    niveauHabitationList: ItemList[];
    typeChantierList: ItemList[];
    puissanceCompteurList: ItemList[];
    generateurList: ItemList[];
    conduitList: ItemList[];
    resistanceList: ItemList[];
    toitureList: ItemList[];
    couleurProfileList: ItemList[];
    couleurFacadeList: ItemList[];
    puissancePoeleList: ItemList[];
    zoneInstallationList: ItemList[];
    typeOrigineList: ItemList[];
}

export default PgList;